Extend and refactor acceptance tests
