+++ test/acceptance/steps/given.js
@@ -1,0 +1,6 @@
+const { Given } = require('@wdio/cucumber-framework');
+const { browser } = require('@wdio/globals')
+
+Given(/^the user navigates to "([^"]*)?"$/, async (page) => {
+    await browser.url(page);
+});
